Algotrading_Crypto_Alpaca/requirements.txt
alpaca-py==0.20.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
flask==2.3.2
pyinstaller==5.13.0
